+++ library/src/test/java/uk/co/samuelwall/materialtaptargetprompt/BuilderUnitTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 Samuel Wall
+ * Copyright (C) 2017-2018 Samuel Wall
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,12 +16,12 @@
 
 package uk.co.samuelwall.materialtaptargetprompt;
 
-
 import android.app.Activity;
 import android.app.Dialog;
-import android.app.DialogFragment;
-import android.app.Fragment;
+import android.support.v4.app.DialogFragment;
+import android.support.v4.app.Fragment;
 
+import android.view.ViewGroup;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.Robolectric;
@@ -32,6 +32,7 @@ import static junit.framework.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
+import static org.robolectric.shadows.support.v4.SupportFragmentTestUtil.startFragment;
 
 @RunWith(RobolectricTestRunner.class)
 @Config(constants = uk.co.samuelwall.materialtaptargetprompt.BuildConfig.class, sdk = 22)
@@ -40,7 +41,7 @@ public class BuilderUnitTest
     @Test
     public void testBuilder_Fragment()
     {
-        final Fragment fragment = Robolectric.buildFragment(Fragment.class).create().get();
+        final android.app.Fragment fragment = Robolectric.buildFragment(android.app.Fragment.class).create().get();
         final MaterialTapTargetPrompt.Builder builder = new MaterialTapTargetPrompt.Builder(fragment);
         assertTrue(builder.getResourceFinder() instanceof ActivityResourceFinder);
     }
@@ -48,7 +49,7 @@ public class BuilderUnitTest
     @Test
     public void testBuilder_Fragment_Resource()
     {
-        final Fragment fragment = Robolectric.buildFragment(Fragment.class).create().get();
+        final android.app.Fragment fragment = Robolectric.buildFragment(android.app.Fragment.class).create().get();
         final MaterialTapTargetPrompt.Builder builder = new MaterialTapTargetPrompt.Builder(fragment, 0);
         assertTrue(builder.getResourceFinder() instanceof ActivityResourceFinder);
     }
@@ -56,7 +57,7 @@ public class BuilderUnitTest
     @Test
     public void testBuilder_DialogFragment()
     {
-        final DialogFragment dialogFragment = spy(Robolectric.buildFragment(DialogFragment.class).create().get());
+        final android.app.DialogFragment dialogFragment = spy(Robolectric.buildFragment(android.app.DialogFragment.class).create().get());
         final Activity activity = Robolectric.buildActivity(Activity.class).create().get();
         final Dialog dialog = mock(Dialog.class);
         when(dialogFragment.getDialog()).thenReturn(dialog);
@@ -83,5 +84,43 @@ public class BuilderUnitTest
         final Activity activity = Robolectric.buildActivity(Activity.class).create().get();
         final MaterialTapTargetPrompt.Builder builder = new MaterialTapTargetPrompt.Builder(activity);
         assertTrue(builder.getResourceFinder() instanceof ActivityResourceFinder);
+    }
+
+    @Test
+    public void testBuilder_SupportFragment()
+    {
+        final Fragment fragment = spy(new Fragment());
+        final ViewGroup view = mock(ViewGroup.class);
+        when(fragment.getView()).thenReturn(view);
+        startFragment(fragment);
+        final MaterialTapTargetPrompt.Builder builder = new MaterialTapTargetPrompt.Builder(fragment);
+        assertTrue(builder.getResourceFinder() instanceof SupportFragmentResourceFinder);
+    }
+
+    @Test
+    public void testBuilder_SupportFragment_Resource()
+    {
+        final Fragment fragment = spy(new Fragment());
+        final ViewGroup view = mock(ViewGroup.class);
+        when(fragment.getView()).thenReturn(view);
+        startFragment(fragment);
+        final MaterialTapTargetPrompt.Builder builder = new MaterialTapTargetPrompt.Builder(fragment, 0);
+        assertTrue(builder.getResourceFinder() instanceof SupportFragmentResourceFinder);
+    }
+
+    @Test
+    public void testBuilder_SupportDialogFragment()
+    {
+        final DialogFragment dialogFragment = spy(new DialogFragment());
+        startFragment(dialogFragment);
+        final ViewGroup view = mock(ViewGroup.class);
+        when(dialogFragment.getView()).thenReturn(view);
+        final Activity activity = Robolectric.buildActivity(Activity.class).create().get();
+        final Dialog dialog = mock(Dialog.class);
+        when(dialogFragment.getDialog()).thenReturn(dialog);
+        when(dialog.getOwnerActivity()).thenReturn(activity);
+        when(dialog.findViewById(android.R.id.content)).thenReturn(activity.findViewById(android.R.id.content));
+        final MaterialTapTargetPrompt.Builder builder = new MaterialTapTargetPrompt.Builder(dialogFragment);
+        assertTrue(builder.getResourceFinder() instanceof SupportFragmentResourceFinder);
     }
 }